Large improvements for the detection of cross-compilation on Mac OS X systems.
The tests for different architecture support now follows the ideas discussed in the post
http://thread.gmane.org/gmane.science.nmr.relax.devel/5785/focus=5820.

In summary, for each architecture a simple C file is created, compiled with 'gcc -arch xyz', and the
resultant binary file tested.  To support 64-bit compilation on 32-bit systems, all previously
successful architectures are also included in the gcc command.

The change allows the 'ppc64' architecture to be reintroduced.



git-svn-id: svn://svn.gna.org/svn/relax/trunk@23638 b7916896-f9f9-0310-9fe5-b3996d8957d5