Hack to permanently eliminate the ^[[?1034h escape code being produced on Linux systems.
This is produced by importing the readline module.  The escape code will be sent to STDOUT every
time relax is executed, so it will be present in all log files.  The problem is the TERM
environmental variable being set to 'xterm'.  The hack simply sets TERM to an empty string.



git-svn-id: svn://svn.gna.org/svn/relax/trunk@24257 b7916896-f9f9-0310-9fe5-b3996d8957d5