Fix for the software verification tests.
The recent expansion and improvements of the REMARK records created by the internal structural
object PDB writing method imported the relax version to place this information into the PDB files.
However this breaks the relax library design, as shown by the verification tests.  Instead the relax
version information is being taken from the lib.structure.internal.object.RELAX_VERSION variable.
This defaults to None, however the version module now sets this variable directly when it is
imported so that it is always set to the current relax version when running relax.



git-svn-id: svn://svn.gna.org/svn/relax/trunk@24386 b7916896-f9f9-0310-9fe5-b3996d8957d5