Fix for the breakage of the relax help system.
This was reported at http://thread.gmane.org/gmane.science.nmr.relax.devel/6481.

The problem was that the TERM environmental variable was turned off to avoid the GNU readline
library on Linux systems emitting the ^[[?1034h escape code.  See the message at
http://thread.gmane.org/gmane.science.nmr.relax.devel/6481/focus=6489 for more details.  However the
Python help system obviously requires this environmental variable.  Now only if the TERM variable is
set to 'xterm' will it be reset, and to 'linux' instead of the blank string ''.



git-svn-id: svn://svn.gna.org/svn/relax/trunk@24412 b7916896-f9f9-0310-9fe5-b3996d8957d5