Reverted r24024, the deletion of the Relax_disp.test_hansen_cpmg_data_to_lm63_3site system test.
See the thread at http://thread.gmane.org/gmane.science.nmr.relax.scm/21774/focus=6300 for the
request that this deletion be reverted.  This is the only system test for the 'LM63 3-site'
dispersion model using real data.  Having this test allows for better coverage of the code.

The command used was:
svn merge -r24024:r24023 .

.....
  r24024 | tlinnet | 2014-06-17 12:32:34 +0200 (Tue, 17 Jun 2014) | 4 lines
  Changed paths:
     M /trunk/test_suite/system_tests/relax_disp.py
  
  Removed systemtest test_hansen_cpmg_data_to_lm63_3site.
  
  This was a temporary implementation and has been replaced with systemtest:
  Relax_disp.test_lm63_3site_synthetic
.....



git-svn-id: svn://svn.gna.org/svn/relax/trunk@24739 b7916896-f9f9-0310-9fe5-b3996d8957d5