Changed the behaviour of the pipe_control.structure.mass.pipe_centre_of_mass() function.
This function returns the CoM and optionally the mass of the structural data loaded into the current
data pipe.  However it was matching the structural data to the molecule-residue-spin data structure
and skipping spins that were deselected.  This illogical deselection part has been eliminated, as
spins can be deselected for various analysis purposes and this should not change the CoM.  The
deletion also significantly speeds up the function.



git-svn-id: svn://svn.gna.org/svn/relax/trunk@24837 b7916896-f9f9-0310-9fe5-b3996d8957d5