Fixes for the relaxation dispersion loop_parameters() function.
The R1, R20, R20A, R20B (and R1rho_prime, R1rho_primeA, R1rho_primeB) parameters are now checked for
in each spin container rather than just the first of the cluster.  This should make no difference as
all spins should have the same model and parameters, but it might be a source of bugs in the future.



git-svn-id: svn://svn.gna.org/svn/relax/trunk@25134 b7916896-f9f9-0310-9fe5-b3996d8957d5