Converted the relaxation dispersion target function class to the unified 'No Rex' model design.
On top of removing all references to MODEL_NOREX_R1RHO, the aliasing of self.func now checks the
experiment type list to determine which target function to use.  This is not an ideal solution and
will not handle mixed CPMG and R1rho experiment, however neither will the target functions yet.

The creation of the off-resonance data structures has also been modified so that they are now R1rho
independent.  This allows the structures to be properly created while at the same time enabling this
code to be compatible with off-resonance CPMG data in the future.

git-svn-id: svn://svn.gna.org/svn/relax/trunk@25166 b7916896-f9f9-0310-9fe5-b3996d8957d5