Added an optimisation script for the test_suite/shared_data/diffusion_tensor/ellipsoid relaxation data.
This is to help catch bug #22502 (https://gna.org/bugs/?22502), the geometric prolate diffusion
representation does not align with axis in PDB, as reported by Martin Ballaschk
(https://gna.org/users/mab).

The PDB files of the optimised tensor demonstrate exactly the same problem as seen in the files
attached to the bug report.  The oblate and spherical diffusion tensor representations match that of
the ellipsoid.  But the prolate axis and tensor orientation are both different from the ellipsoid
as well as themselves.



git-svn-id: svn://svn.gna.org/svn/relax/trunk@25204 b7916896-f9f9-0310-9fe5-b3996d8957d5