Changed the optimisation description in the relaxation curve-fitting chapter of the manual.
The script example has been converted to match the sample script, replacing the Nelder-Mead simplex
algorithm with Newton optimisation, and removing the argument turning diagonal scaling off.  All the
text about only the simplex algorithm being supported due to the missing gradients and Hessians in
the C module have been deleted.  The text that linear constraints are not supported has also been
removed - but this was fixed when the logarithmic barrier constraint algorithm was added to minfx.



git-svn-id: svn://svn.gna.org/svn/relax/trunk@25336 b7916896-f9f9-0310-9fe5-b3996d8957d5