Speed up of the target_functions.relax_fit C module.
The variances are now precalculated in the setup() function from the errors, so that the use of the
square() function is minimised.  The chi-squared equation, gradient, and Hessian functions now
accept the variance rather than standard deviation argument and hence the squaring of errors has
been removed.  This avoids a lot of duplicated maths operations.



git-svn-id: svn://svn.gna.org/svn/relax/trunk@25414 b7916896-f9f9-0310-9fe5-b3996d8957d5