The dispersion get_param_values() API method now calls the r1_setup() function.
This relates to bug #22541 (https://gna.org/bugs/?22541), the R1 fit flag does not work in the GUI.

This is to make sure that the parameters are correctly set up prior to obtaining all parameter
values.  The R1 parameter is dynamic hence r1_setup() needs to be called at any point model
parameters are accessed, as the R1 parameter can be turned on or off at any time with the
relax_disp.r1_fit user function.



git-svn-id: svn://svn.gna.org/svn/relax/trunk@25500 b7916896-f9f9-0310-9fe5-b3996d8957d5