The model argument in the structure.delete user function is now operational.
In the internal object, it has two functions.  When the atom_id argument is none, then new
ModelList.delete_model() function is being called to remove the entire model from the list of
structural models.  When the atom_id argument is supplied, then only the corresponding atoms in the
given model will be deleted.



git-svn-id: svn://svn.gna.org/svn/relax/trunk@25773 b7916896-f9f9-0310-9fe5-b3996d8957d5